thor: change `getCpuCount` to return a `size_t` instead of `int`

+++ kernel/thor/arch/x86/cpu.cpp
@@ -535,7 +535,7 @@ CpuData *getCpuData(size_t k) {
 	return (*allCpuContexts)[k];
 }
 
-int getCpuCount() {
+size_t getCpuCount() {
 	return allCpuContexts->size();
 }
